pyproject.toml
[tool.poetry]
name = "chat-with-your-data-solution-accelerator"
version = "0.1.0"
description = "Chat with your data solution accelerator"
authors = []
readme = "README.md"
packages = [
    { include = "code" }
]

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "1.19.0"
streamlit = "1.33.0"
python-dotenv = "1.0.1"
azure-ai-formrecognizer = "3.3.3"
azure-storage-blob = "12.19.1"
azure-identity = "1.16.0"
flask = {extras = ["async"], version = "^3.0.3"}
openai = "1.24.0"
langchain = "0.1.16"
langchain-community = "0.0.34"
langchain-openai = "0.1.4"
requests = "2.31.0"
tiktoken = "0.6.0"
azure-storage-queue = "12.9.0"
beautifulsoup4 = "4.12.3"
fake-useragent = "1.5.1"
chardet = "5.2.0"
azure-search-documents = "11.6.0b1"
azure-ai-contentsafety = "1.0.0"
python-docx = "1.1.0"
azure-keyvault-secrets = "4.8.0"
pandas = "2.2.2"
azure-monitor-opentelemetry = "^1.4.1"
opentelemetry-instrumentation-httpx = "^0.45b0"
pillow = "10.3.0"
azure-mgmt-cognitiveservices = "^13.5.0"
jsonschema = "^4.21.1"
semantic-kernel = {version = "^0.9.6b1", python = "<3.13"}
azure-ai-documentintelligence = "^1.0.0b3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "5.0.0"
flake8 = "7.0.0"
black = "24.4.2"
pre-commit = "3.7.0"
pytest_httpserver = "1.0.10"
trustme = "1.1.0"
jupyter = "1.0.0"
pytest-asyncio = "^0.23.6"

[tool.coverage.run]
omit = [
    "**/tests/*",
    "**/dist/*",
]

[tool.coverage.report]
skip_empty = true
include_namespace_packages = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
